+++ packages/frontend/.storybook/main.ts
@@ -1,25 +1,12 @@
 import type { StorybookConfig } from '@storybook/react-vite';
 
-import { dirname, join } from 'path';
-
-/**
- * This function is used to resolve the absolute path of a package.
- * It is needed in projects that use Yarn PnP or are set up within a monorepo.
- */
-function getAbsolutePath(value: string): any {
-    return dirname(require.resolve(join(value, 'package.json')));
-}
 const config: StorybookConfig = {
     stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
     addons: [
-        getAbsolutePath('@storybook/addon-essentials'),
-        getAbsolutePath('@storybook/addon-onboarding'),
-        getAbsolutePath('@chromatic-com/storybook'),
-        getAbsolutePath('@storybook/addon-interactions'),
+        '@storybook/addon-docs',
+        '@storybook/addon-onboarding',
+        '@chromatic-com/storybook',
     ],
-    framework: {
-        name: getAbsolutePath('@storybook/react-vite'),
-        options: {},
-    },
+    framework: '@storybook/react-vite',
 };
 export default config;